Modularized Test Runner.
Added Module tests in emotly. Added a single file class for each TestCase.
Added a test suites in Run Tests.
